requirements.txt
PyMuPDF>=1.22.5
pytesseract>=0.3.10
opencv-python>=4.8.0.74
numpy>=1.25.0
pandas>=2.1.0
camelot-py[cv]>=0.10.1
tk
